Completely refactor sd.c to follow correct protocol
Create a new function sd_send_cmd() to generalize the calls to the SPI
driver which makes it easier to call from the other functions and
cleaner and easier to read.

Refactor the init() function to follow the proper protocol for
initializing a newer SD card with fallbacks for old-card compatibility
just in case.

Add a lot of commenting to explain what is happening where and what the
proper procedure is to start up. Add links to relevant websites where
information can be found and add explanations and concerns for working
with the SD card in SPI mode in comments at the top of the file.