Convert SD Card Driver to service-style module
Convert the interface for writing to the SD card to the style of the
other avionics components where the SD card will write a block every
iteration of main_loop() except when the card is busy.

There are still a few decisions to be made such as where the data buffer
to be written will live/how the card will access this data and whether
or not the card should be written to on a timer basis or as it is
currently.

Additional cleanup throughout driver code.