Add one more state and break up firing of pins
Both ematch pins were being driven at the same time. This would result
in both the main and drogue firing at the same time. The extra added
states in this and the previous commit will break up the firing of
the ematches.

A maximum altitude requirement was also added for deployment, and sample
values for these added.

Finally, the pins used to fire the ematches have been renamed to make
what they do more clear, though confirmation is needed on if they are
the correct ones.